refac: merge the transformation modules
