+++ source/bronze/src/opengeh_bronze/application/streams/notify_transactions_persisted_stream.py
@@ -1,7 +1,6 @@
 from pyspark.sql import DataFrame
 
 import opengeh_bronze.application.config.spark_session as spark_session
-import opengeh_bronze.domain.transformations.bronze_measurements_transformation as bronze_measurements_transformation
 import opengeh_bronze.domain.transformations.notify_transactions_persisted_events_transformation as notify_transactions_persisted_events_transformation
 import opengeh_bronze.infrastructure.streams.writer as writer
 from opengeh_bronze.infrastructure.streams.bronze_repository import BronzeRepository
@@ -15,9 +14,8 @@ def notify() -> None:
     writer.write_stream(bronze_stream, "query_name", options, notify_transactions_persisted)
 
 
-def notify_transactions_persisted(bronze_measurements: DataFrame, batch_id: int) -> None:
-    unpackaged_measurements = bronze_measurements_transformation.transform(bronze_measurements)
+def notify_transactions_persisted(submitted_transactions: DataFrame, batch_id: int) -> None:
     notify_transactions_persisted_events = notify_transactions_persisted_events_transformation.transform(
-        unpackaged_measurements
+        submitted_transactions
     )
     KafkaStream().write_stream(notify_transactions_persisted_events)
