+++ source/bronze/src/opengeh_bronze/infrastructure/helpers/path_helper.py
@@ -1,6 +1,5 @@
-from importlib.abc import Traversable
 from importlib.resources import files
 
 
-def get_protobuf_descriptor_path(file_name: str) -> Traversable:
-    return files("opengeh_bronze.infrastructure.contracts.assets").joinpath(file_name)
+def get_protobuf_descriptor_path(file_name: str) -> str:
+    return str(files("opengeh_bronze.infrastructure.contracts.assets").joinpath(file_name))
